requirements.txt
sqlalchemy-utils
python-dotenv
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
gunicorn==25.0.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
PyMySQL==1.1.2
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.5
